requirements.txt
requests==2.32.3
dotenv==0.9.9
python-telegram-bot==13.15
standard-imghdr==3.13.0
urllib3==1.26.15